+++ pattern/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:asp/asp.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
@@ -6,12 +5,5 @@ import 'app/app_module.dart';
 import 'app/app_widget.dart';
 
 void main() {
-
-  // * Support methods
-  Modular.setInitialRoute('/home/');
-
-  // * Start
-  return runApp(
-    ModularApp(module: AppModule(), child: const RxRoot(child: AppWidget())),
-  );
+  runApp(ModularApp(module: AppModule(), child: const AppWidget()));
 }